add: changed to use the new QvMsgBus scheme.
