+++ Qv2ray.pro
@@ -138,7 +138,6 @@ Qv2rayAddSource(base, models, QvConfigIdentifier, hpp)
 Qv2rayAddSource(base, models, QvSafeType, hpp)
 Qv2rayAddSource(base, models, QvRuntimeConfig, hpp)
 Qv2rayAddSource(base, models, QvStartupConfig, hpp)
-Qv2rayAddSource(base, messaging, QvGlobalMessageBus, cpp, hpp)
 Qv2rayAddSource(common, _, CommandArgs, cpp, hpp)
 Qv2rayAddSource(common, _, HTTPRequestHelper, cpp, hpp)
 Qv2rayAddSource(common, _, LogHighlighter, cpp, hpp)
@@ -163,6 +162,7 @@ Qv2rayAddSource(core, connection, Serialization, cpp, hpp)
 Qv2rayAddSource(core, _, CoreUtils, cpp, hpp)
 Qv2rayAddSource(core, kernel, KernelInteractions, cpp, hpp)
 Qv2rayAddSource(core, kernel, APIBackend, cpp, hpp)
+Qv2rayAddSource(ui, uibase, QvMessageBus, cpp, hpp)
 Qv2rayAddSource(ui, editors, w_InboundEditor, cpp, hpp, ui)
 Qv2rayAddSource(ui, editors, w_JsonEditor, cpp, hpp, ui)
 Qv2rayAddSource(ui, editors, w_OutboundEditor, cpp, hpp, ui)
