+++ src/ui/editors/w_JsonEditor.cpp
@@ -5,7 +5,7 @@
 JsonEditor::JsonEditor(QJsonObject rootObject, QWidget *parent) :
     QDialog(parent)
 {
-    QvMessageBusConnect(JsonEditor);
+    QvMsgBusSlot(QvMsgBusImplDefault)
     setupUi(this);
     original = rootObject;
     final = rootObject;
@@ -23,8 +23,6 @@ JsonEditor::JsonEditor(QJsonObject rootObject, QWidget *parent) :
     jsonTree->expandAll();
     jsonTree->resizeColumnToContents(0);
 }
-
-QvMessageBusSlotImplDefault(JsonEditor)
 
 QJsonObject JsonEditor::OpenEditor()
 {
