+++ src/ui/editors/w_JsonEditor.hpp
@@ -6,6 +6,7 @@
 #include "base/Qv2rayBase.hpp"
 #include "ui_w_JsonEditor.h"
 
+
 class JsonEditor : public QDialog, private Ui::JsonEditor
 {
         Q_OBJECT
@@ -14,8 +15,6 @@ class JsonEditor : public QDialog, private Ui::JsonEditor
         explicit JsonEditor(QJsonObject rootObject, QWidget *parent = nullptr);
         ~JsonEditor();
         QJsonObject OpenEditor();
-    public slots:
-        QvMessageBusSlotHeader
 
     private slots:
         void on_jsonEditor_textChanged();
