+++ src/ui/editors/w_RoutesEditor.cpp
@@ -71,7 +71,7 @@ void RouteEditor::SetupNodeWidget()
 
 RouteEditor::RouteEditor(QJsonObject connection, QWidget *parent) : QDialog(parent), root(connection), original(connection)
 {
-    QvMessageBusConnect(RouteEditor);
+    QvMsgBusSlot(QvMsgBusImplDefault)
     setupUi(this);
     isLoading = true;
     setWindowFlags(windowFlags() | Qt::WindowMaximizeButtonHint);
@@ -120,8 +120,6 @@ RouteEditor::RouteEditor(QJsonObject connection, QWidget *parent) : QDialog(pare
 
     isLoading = false;
 }
-
-QvMessageBusSlotImplDefault(RouteEditor)
 
 void RouteEditor::onNodeClicked(Node &n)
 {
