+++ src/ui/w_ExportConfig.hpp
@@ -4,6 +4,7 @@
 #include "base/Qv2rayBase.hpp"
 #include "3rdparty/qzxing/src/QZXing.h"
 
+
 class ConfigExporter : public QDialog, private Ui::ExportConfigWindow
 {
         Q_OBJECT
@@ -12,8 +13,6 @@ class ConfigExporter : public QDialog, private Ui::ExportConfigWindow
         explicit ConfigExporter(const CONFIGROOT &root, const ConnectionIdentifier &alias, QWidget *parent = nullptr);
         ~ConfigExporter();
         void OpenExport();
-    public slots:
-        QvMessageBusSlotHeader
     protected:
         void changeEvent(QEvent *e);
     private slots:
