+++ src/ui/w_ImportConfig.cpp
@@ -24,11 +24,9 @@ ImportConfigWindow::ImportConfigWindow(QWidget *parent)
 {
     setupUi(this);
     nameTxt->setText(QDateTime::currentDateTime().toString("MMdd_hhmm"));
-    QvMessageBusConnect(ImportConfigWindow);
+    QvMsgBusSlot(QvMsgBusImplDefault)
     RESTORE_RUNTIME_CONFIG(screenShotHideQv2ray, hideQv2rayCB->setChecked)
 }
-
-QvMessageBusSlotImplDefault(ImportConfigWindow)
 
 ImportConfigWindow::~ImportConfigWindow()
 {
@@ -55,9 +53,10 @@ void ImportConfigWindow::on_qrFromScreenBtn_clicked()
     bool hideQv2ray = hideQv2rayCB->isChecked();
 
     if (hideQv2ray) {
-        messageBus.EmitGlobalSignal(QvMessage::HIDE_WINDOWS);
+        messageBus.EmitGlobalSignal(QvMessage::MINIMIZE_WINDOWS);
     }
 
+    QApplication::processEvents();
     QThread::msleep(static_cast<ulong>(doubleSpinBox->value() * 1000));
     auto w = new ScreenShotWindow();
     auto pix = w->DoScreenShot();
@@ -66,7 +65,7 @@ void ImportConfigWindow::on_qrFromScreenBtn_clicked()
     delete w;
 
     if (hideQv2ray) {
-        messageBus.EmitGlobalSignal(QvMessage::SHOW_WINDOWS);
+        messageBus.EmitGlobalSignal(QvMessage::RESTORE_WINDOWS);
         //ShowAllGlobalWindow();
     }
