+++ src/ui/w_ImportConfig.ui
@@ -2,6 +2,9 @@
 <ui version="4.0">
  <class>ImportConfigWindow</class>
  <widget class="QDialog" name="ImportConfigWindow">
+  <property name="windowModality">
+   <enum>Qt::ApplicationModal</enum>
+  </property>
   <property name="geometry">
    <rect>
     <x>0</x>
