+++ src/ui/w_MainWindow.cpp
@@ -79,8 +79,6 @@
 
 MainWindow *MainWindow::mwInstance = nullptr;
 
-QvMessageBusSlotImplDefault(MainWindow)
-
 MainWindow::MainWindow(QWidget *parent):
     QMainWindow(parent), vinstance(),
     hTray(new QSystemTrayIcon(this)), vCoreLogHighlighter(), qvAppLogHighlighter()
@@ -96,7 +94,7 @@ MainWindow::MainWindow(QWidget *parent):
     });
     //
     setupUi(this);
-    QvMessageBusConnect(MainWindow);
+    QvMsgBusSlot(QvMsgBusImplDefault)
     //
     // Two browsers
     logTextBrowsers.append(new QTextBrowser());
