+++ src/ui/w_SubscriptionManager.hpp
@@ -3,8 +3,9 @@
 #include <QDialog>
 #include "base/Qv2rayBase.hpp"
 #include "common/HTTPRequestHelper.hpp"
-
 #include "ui_w_SubscriptionManager.h"
+
+
 class SubscribeEditor : public QDialog, private Ui::w_SubscribeEditor
 {
         Q_OBJECT
